perf: 修改账号配置
