+++ apps/accounts/models/__init__.py
@@ -1,3 +1,5 @@
 from .account import *
 from .automations import *
 from .base import *
+from .template import *
+from .virtual import *
